as: Refactor request creation
